requirements.txt
azure-functions==1.18.0
azure-cosmosdb-table==1.0.6